+++ CHANGELOG.md
@@ -15,8 +15,7 @@ This document outlines major changes between releases.
 ### Changed
 - Go 1.23+ is required to build now (#235)
 - Maximum search limit shrunk from 10000 to 1000 (#276)
-- Get and Head by attributes, now sorts result by `Timestamp` attribute. The gate searches only 1000 elements
-and skips others (#276)
+- Attribute-based GET and HEAD now return the latest (wrt `Timestamp` attribute) object from the first 1000 search results (#276)
 
 ### Removed